+++ src/main/java/com/example/planeo_back/domain/ports/BalanceRepository.java
@@ -1,8 +1,10 @@
 package com.example.planeo_back.domain.ports;
 import com.example.planeo_back.domain.models.balance.BalanceDomain;
 
+import java.math.BigDecimal;
+
 public interface BalanceRepository extends IGenericCrudRepository<BalanceDomain> {
     BalanceDomain findBalanceByUsername(String username);
-    void decreaseCurrentBalance(String username, double amount);
+    void decreaseCurrentBalance(String username, BigDecimal amount);
     boolean balanceExistForUser(String username);
 }
