+++ src/main/java/com/example/planeo_back/web/DTO/BalanceResponseDTO.java
@@ -1,9 +1,11 @@
 package com.example.planeo_back.web.DTO;
 
+import java.math.BigDecimal;
+
 public record BalanceResponseDTO(
      Long id,
-     Double currentBalance,
-     Double futureBalance,
-     Double pendingExpense
+     BigDecimal currentBalance,
+     BigDecimal futureBalance,
+     BigDecimal pendingExpense
     ) {}
